mesh_api/requirements.txt
bleak==0.21.1
bleak-winrt==1.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==3.4.8
dotmap==1.3.30
enum-compat==0.0.3
Flask==2.2.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
meshtastic==1.2.54
packaging==24.1
pexpect==4.9.0
print-color==0.4.6
protobuf==5.27.2
ptyprocess==0.7.0
pycparser==2.22
pygatt==4.0.5
pyparsing==3.1.2
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
PyYAML==6.0.1
requests==2.32.3
tabulate==0.9.0
timeago==1.0.16
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
pytest==7.7.2


